Extend nodekey and attrstring when undecrypted node is serialize

+++ src/megaclient.cpp
@@ -19956,13 +19956,13 @@ Node *NodeManager::unserializeNode(const std::string *d, bool fromOldCache)
     if (encrypted)
     {
         // Have we encoded the node key data's length?
-        if (ptr + sizeof(unsigned short) > end)
+        if (ptr + sizeof(uint32_t) > end)
         {
             mNodes.erase(nodePosition);
             return nullptr;
         }
 
-        auto length = MemAccess::get<unsigned short>(ptr);
+        auto length = MemAccess::get<uint32_t>(ptr);
         ptr += sizeof(length);
 
         // Have we encoded the node key data?
@@ -19976,13 +19976,13 @@ Node *NodeManager::unserializeNode(const std::string *d, bool fromOldCache)
         ptr += length;
 
         // Have we encoded the length of the attribute string?
-        if (ptr + sizeof(unsigned short) > end)
+        if (ptr + sizeof(uint32_t) > end)
         {
             mNodes.erase(nodePosition);
             return nullptr;
         }
 
-        length = MemAccess::get<unsigned short>(ptr);
+        length = MemAccess::get<uint32_t>(ptr);
         ptr += sizeof(length);
 
         // Have we encoded the attribute string?
